🎨 custom hookを用いてrecommendのロジックを切り離した
